requirements/common.txt

Django

django_csp
django-cors-headers

psycopg2

redis
celery[redis]
django-redis-sessions
django-redis-cache
hiredis

flake8

djangorestframework
markdown
djangorestframework-jwt

ldap3
pytz

oauth2client
google-api-python-client

pylint

requests


